Fix nightly and improve merge error message
The nightly bots are all failing because a nightly build emits a .cargo/config
which uses `[target]` for some overrides, but a test was creating invalid
configuration for `[target]`, causing the test error message to change.

This commit fixes the test by using a different error message (which should not
conflict). It also improves the error message for "configuration could not be
merged" to include more contextual information about what just happened.